Align tests with package structure and avoid warnings
